requirements.txt
flask
torch
torchvision
pillow
numpy
opencv-python
pandas
gunicorn
